refactor(backend): extract container and port routers (phase 4c-4)
Move the six /api/containers/* and /api/ports/in-use endpoints out of
index.ts. Handlers moved verbatim. routes/containers.ts exports two
routers:
- containersRouter (mounted at /api/containers): list, stream logs,
  start, stop, restart.
- portsRouter (mounted at /api/ports): /in-use host port inventory.

Removes the now-unused requireAdmin import from index.ts. Middleware
chains, response shapes, and error messages are all preserved.

index.ts drops from 1437 to 1364 lines. Remaining inline groups:
stacks and nodes.
